chore: make a util C library for Lean FFI
We just copied the code, but now we need to link multiple
libraries (e.g. archive and z)
