+++ KLR/Util/Archive/lake-manifest.json
@@ -1,7 +1,14 @@
 {"version": "1.1.0",
  "packagesDir": ".lake/packages",
  "packages":
- [{"url": "https://github.com/leanprover/lean4-cli.git",
+ [{"type": "path",
+   "scope": "",
+   "name": "FFIUtil",
+   "manifestFile": "lake-manifest.json",
+   "inherited": false,
+   "dir": "./../FFIUtil",
+   "configFile": "lakefile.lean"},
+  {"url": "https://github.com/leanprover/lean4-cli.git",
    "type": "git",
    "subDir": null,
    "scope": "",
