+++ KLR/Util/FFIUtil/include/lean_util.h
@@ -1,0 +1,3 @@
+#include <lean/lean.h>
+
+lean_object* io_err(const char* fmt, ...);
