+++ KLR/Util/FFIUtil/lake-manifest.json
@@ -1,0 +1,5 @@
+{"version": "1.1.0",
+ "packagesDir": ".lake/packages",
+ "packages": [],
+ "name": "ffi_util",
+ "lakeDir": ".lake"}
