+++ KLR/Util/Gzip/lakefile.lean
@@ -20,7 +20,7 @@ def moreLinkArgs :=
     ]
 
 def moreWeakArgs :=
-  let all := #[]
+  let all := #["-I", "include"]
   if System.Platform.isOSX then
     all ++ #[
       "-I", "/opt/homebrew/opt/zlib/include",
@@ -43,7 +43,8 @@ lean_exe gzip where
 target lean_gzip.o pkg : FilePath := do
   let oFile := pkg.buildDir / "lean_gzip.o"
   let srcJob ← inputTextFile <| pkg.dir / "lean_gzip.c"
-  let weakArgs := #["-I", (← getLeanIncludeDir).toString] ++ moreWeakArgs
+  let ffiutil := pkg.dir / ".." / "FFIUtil" / "include"
+  let weakArgs := #["-I", ffiutil.toString, "-I", (← getLeanIncludeDir).toString] ++ moreWeakArgs
   let compilerFlags := #["-fPIC", "-Werror"]
   buildO oFile srcJob weakArgs compilerFlags "cc" getLeanTrace
 
@@ -54,3 +55,5 @@ extern_lib liblean_gzip pkg := do
 
 require Cli from git
   "https://github.com/leanprover/lean4-cli.git" @ "v4.19.0"
+
+require FFIUtil from "../FFIUtil"
