+++ lake-manifest.json
@@ -81,6 +81,13 @@
    "manifestFile": "lake-manifest.json",
    "inputRev": "v4.19.0",
    "inherited": true,
-   "configFile": "lakefile.toml"}],
+   "configFile": "lakefile.toml"},
+  {"type": "path",
+   "scope": "",
+   "name": "FFIUtil",
+   "manifestFile": "lake-manifest.json",
+   "inherited": true,
+   "dir": "./KLR/Util/Gzip/./../FFIUtil",
+   "configFile": "lakefile.lean"}],
  "name": "KLR",
  "lakeDir": ".lake"}
